feat: config WebSocket and update model message, room

+++ src/main/java/iuh/fit/fashionecommercewebsitebackend/repositories/RoomChatRepository.java
@@ -3,5 +3,5 @@ package iuh.fit.fashionecommercewebsitebackend.repositories;
 import iuh.fit.fashionecommercewebsitebackend.models.RoomChat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomChatRepository extends JpaRepository<RoomChat, Integer> {
+public interface RoomChatRepository extends JpaRepository<RoomChat, Long> {
 }
